manhattanBeach_scraper/requirements.txt
requests
bs4
pandas==0.24.2
selenium==3.141.0
pymongo==3.10.1
dnspython=1.16
argparse
py3_validate_email